\name{nloptwrap-package}
\alias{nloptwrap-package}
\alias{nloptwrap}
\docType{package}
\title{
  Wrapper for NLOPT functions
}
\description{
  Wraps optimization functions in `nloptr' to make them more accessible
  and easier to use for casual optimization tasks in R. 
}
\details{
  \tabular{ll}{
  Package: \tab nloptwrap\cr
  Type: \tab Package\cr
  Version: \tab 0.4-4\cr
  Date: \tab 2013-02-06\cr
  License: \tab LGPL\cr
  }
}
\author{
Hans W. Borchers

Maintainer: HwB <hwborchers@googlemail.com>
}
\references{
  \url{http://ab-initio.mit.edu/wiki/index.php/NLopt_Algorithms}
}
